client/requirements.txt
pystray
pillow